feat: 增强输出日志的可读性，修正风队测试

+++ src/main/java/xiaor/charas/Chara.java
@@ -26,7 +26,7 @@ public abstract class Chara{
 
     protected SnapShot attackShot = new SnapShot(this);
 
-    protected int life;
+    protected long life;
 
     protected double baseLife;
